fix: Support for BMFF V3 hashing options (#1253)
* Support for BMFF Hash V3
validation of fixedBlockSize and fixedBlockSizes Merkle trees
generation of fixedBlockSize Merkle trees

* Bug fix for case when BMFF exclusion subset is using length with no offset

* Setting to control Merkle tree storage

* Change error message

* More error checking
PR comment change

* Make settings fetch return error is not set

* Add test for no fixed offset

* Fix formatting

* Check for conflicting block types

* Fix error response for competing Merkle block types

* Make spec responses

* Allow 0 length fixed offset

* A little cleanup

* Make sure there is a recognized hash binding

* Fix for some more edge cases

* clippy fix

---------

+++ sdk/src/manifest.rs
@@ -1352,7 +1352,7 @@ impl Manifest {
         if dh.is_err() {
             let mut ph = DataHash::new("jumbf manifest", "sha256");
             for _ in 0..10 {
-                ph.add_exclusion(HashRange::new(0, 2));
+                ph.add_exclusion(HashRange::new(0u64, 2u64));
             }
             self.add_assertion(&ph)?;
         }
@@ -2808,7 +2808,7 @@ pub(crate) mod tests {
         // build manifest to insert in the hole
 
         // create an hash exclusion for the manifest
-        let exclusion = HashRange::new(offset, placeholder.len());
+        let exclusion = HashRange::new(offset as u64, placeholder.len() as u64);
         let exclusions = vec![exclusion];
 
         let mut dh = DataHash::new("source_hash", "sha256");
@@ -2867,7 +2867,7 @@ pub(crate) mod tests {
         // build manifest to insert in the hole
 
         // create an hash exclusion for the manifest
-        let exclusion = HashRange::new(offset, placeholder.len());
+        let exclusion = HashRange::new(offset as u64, placeholder.len() as u64);
         let exclusions = vec![exclusion];
 
         let mut dh = DataHash::new("source_hash", "sha256");
